refactor: organize registration form files, move test to utils, remove unused file (Tue, Jun 17, 2025 12:21 AM)
