+++ src/pages/registration.tsx
@@ -1,4 +1,4 @@
-import RegisterForm from '../components/RegisterForm';
+import RegisterForm from '../components/RegisterForm/RegisterForm';
 
 export default function RegisterPage() {
   return <RegisterForm />;
