+++ src/utils/AddressForm.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi } from 'vitest';
 import { render, screen, fireEvent } from '@testing-library/react';
-import AddressForm from './AddressForm';
+import AddressForm from './RegisterForm/AddressForm';
 
 describe('AddressForm', () => {
   const mockAddress = {
